Move dedup types into storage-api crate
We move the dedup types into the storage-api crate so that we can
implement StorageSerde for them based on the protobuf types defined
in this crate.
