+++ crates/worker/src/partition/shuffle.rs
@@ -13,8 +13,8 @@ use crate::partition::types::OutboxMessageExt;
 use async_channel::{TryRecvError, TrySendError};
 use restate_bifrost::Bifrost;
 use restate_core::cancellation_watcher;
+use restate_storage_api::deduplication_table::DedupInformation;
 use restate_storage_api::outbox_table::OutboxMessage;
-use restate_types::dedup::DedupInformation;
 use restate_types::identifiers::{LeaderEpoch, PartitionId, PartitionKey, WithPartitionKey};
 use restate_types::message::{AckKind, MessageIndex};
 use restate_types::NodeId;
